requirements.txt
fastapi
uvicorn[standard]
google-cloud-speech
google-cloud-storage
google-cloud-firestore
google-cloud-pubsub
google-api-python-client